+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-- Switching a user’s primary payment source now ensures it is also set at the default payment method in Stripe ([#235](https://github.com/craftcms/commerce-stripe/issues/235))
-- Added the `craft\commerce\stripe\services\Customers::EVENT_BEFORE_CREATE_CUSTOMER` event. ([#233](https://github.com/craftcms/commerce-stripe/pull/233))
-- The `craft\commerce\stripe\events\SubscriptionRequestEvent` now has access to the plan being subscribed to. ([#141](https://github.com/craftcms/commerce-stripe/pull/141))
+- Stripe customers’ default payment methods are now kept in sync with Craft users’ primary payment sources. ([#235](https://github.com/craftcms/commerce-stripe/issues/235))
+- Added `craft\commerce\stripe\services\Customers::EVENT_BEFORE_CREATE_CUSTOMER`. ([#233](https://github.com/craftcms/commerce-stripe/pull/233))
+- Added `craft\commerce\stripe\events\SubscriptionRequestEvent::$plan`, which will be set to the plan being subscribed to. ([#141](https://github.com/craftcms/commerce-stripe/pull/141))
 
 ## 3.1.0 - 2022-01-29